fix: log swallowed download errors, dedupe LiteRT compatibility probe
Restore exception logging in ModelDownloadWorker.doWork and
ModelRepository.importLocalModel, extract the duplicated
verifyLiteRtCompatibility into LiteRtCompatibility, and hoist the
legacy-credential migration out of the huggingFaceToken getter.
